Janno

+++ src/main/java/org/example/carrentalservice/billing/api/dto/InvoiceResponse.java
@@ -5,12 +5,16 @@ import lombok.Data;
 
 import java.time.LocalDate;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class InvoiceResponse {
     private Long invoiceId;
     private Long rentalId;
-    private Double amount;
+    //private Double amount;
+
+    private BigDecimal amount;
     private String paymentStatus;
     private LocalDate paymentDate;
 }
